restructured the client
