+++ .github/workflows/mvn-build.yml
@@ -1,0 +1,17 @@
+name: Build via Maven and run tests
+
+on: [pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Java environment
+        uses: actions/setup-java@v4
+        with:
+          java-version: '21'
+          distribution: 'adopt'
+          cache: maven
+      - name: Build with Maven
+        run: mvn verify -PcheckFormat -B
