+++ CODE_OF_CONDUCT.md
@@ -1,3 +1,3 @@
-# Code of Conduct 
+# Code of Conduct
 
 ### View the [Camunda Code of Conduct](https://camunda.com/events/code-conduct/) and find ways to report violations.
