+++ extension/client-java/src/main/java/io/camunda/tasklist/auth/Authentication.java
@@ -1,0 +1,10 @@
+package io.camunda.tasklist.auth;
+
+import java.util.Map;
+
+public interface Authentication {
+
+  Map<String, String> getTokenHeader();
+
+  void resetToken();
+}
