+++ extension/client-java/src/main/java/io/camunda/tasklist/auth/JwtCredential.java
@@ -1,0 +1,5 @@
+package io.camunda.tasklist.auth;
+
+import java.net.URL;
+
+public record JwtCredential(String clientId, String clientSecret, String audience, URL authUrl) {}
