+++ extension/client-java/src/main/java/io/camunda/tasklist/auth/TokenResponse.java
@@ -1,0 +1,51 @@
+package io.camunda.tasklist.auth;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class TokenResponse {
+
+  @JsonProperty("access_token")
+  private String accessToken;
+
+  private String scope;
+
+  @JsonProperty("expires_in")
+  private Integer expiresIn;
+
+  @JsonProperty("token_type")
+  private String tokenType;
+
+  TokenResponse() {}
+
+  public String getAccessToken() {
+    return accessToken;
+  }
+
+  public void setAccessToken(String accessToken) {
+    this.accessToken = accessToken;
+  }
+
+  public String getScope() {
+    return scope;
+  }
+
+  public void setScope(String scope) {
+    this.scope = scope;
+  }
+
+  public Integer getExpiresIn() {
+    return expiresIn;
+  }
+
+  public void setExpiresIn(Integer expiresIn) {
+    this.expiresIn = expiresIn;
+  }
+
+  public String getTokenType() {
+    return tokenType;
+  }
+
+  public void setTokenType(String tokenType) {
+    this.tokenType = tokenType;
+  }
+}
