+++ extension/client-java/src/main/java/io/camunda/tasklist/dto/TaskList.java
