+++ extension/client-java/src/main/java/io/camunda/tasklist/exception/TaskListException.java
@@ -1,0 +1,21 @@
+package io.camunda.tasklist.exception;
+
+public class TaskListException extends Exception {
+
+
+  public TaskListException() {
+    super();
+  }
+
+  public TaskListException(Exception e) {
+    super(e);
+  }
+
+  public TaskListException(String message) {
+    super(message);
+  }
+
+  public TaskListException(String message, Exception e) {
+    super(message, e);
+  }
+}
