+++ extension/client-java/src/main/java/io/camunda/tasklist/util/JsonUtils.java
