+++ extension/client-java/src/main/resources/application.properties
