+++ extension/pom.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>io.camunda</groupId>
+    <artifactId>tasklist-client-root</artifactId>
+    <version>8.6.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>tasklist-client-parent</artifactId>
+  <packaging>pom</packaging>
+  <modules>
+    <module>generated</module>
+    <module>client-java</module>
+  </modules>
+
+</project>
